fix all_types table column

+++ go/test/endtoend/vtgate/queries/misc/schema.sql
@@ -49,7 +49,7 @@ create table all_types
     smallint_unsigned  SMALLINT,
     mediumint_unsigned MEDIUMINT,
     int_unsigned       INT,
-    float_unsigned     FLOAT(10, 2),
+    float_unsigned     FLOAT(10),
     double_unsigned    DOUBLE(16, 2),
     decimal_unsigned   DECIMAL,
     t_date             DATE,
